lib - factor out lib/common and refactor command to be like plugins
