+++ lib/index.js
@@ -1,3 +1,4 @@
 const notes = require('./command.notes')
+const changelog = require('./command.changelog')
 
-module.exports = { notes }
+module.exports = { notes, changelog }
