+++ lib/utils.js
@@ -75,19 +75,15 @@ module.exports.formatNotes = ({ parsed, version, previous }) => {
   return `${notes}\n`
 }
 
-module.exports.writeNotes = ({ notes, version, previous, file }) => {
+module.exports.write = (command = 'notes') => ({ formatted, version, previous, file }) => {
   const invalidFNChars = /[\\/:"*?<>|#%$&{}!+`='" ]+/g
-  const f = file || `release-notes-${previous}-${version}.md`.toLowerCase().replace(invalidFNChars, '-')
-  fs.writeFile(
-    f,
-    notes,
-    (err) => {
-      if (err) {
-        log(`${f} ✖️:`, { severity: 'error' })
-        log(err, { severity: 'error' })
-      } else {
-        log(`${f} ✔️`, { verbose: true })
-      }
-    },
-  )
+  const f = file || `release-${command}-${previous}-${version}.md`.toLowerCase().replace(invalidFNChars, '-')
+  fs.writeFile(f, formatted, (err) => {
+    if (err) {
+      log(`${f} ✖️:`, { severity: 'error' })
+      log(err, { severity: 'error' })
+    } else {
+      log(`${f} ✔️`, { verbose: true })
+    }
+  })
 }
